Fix: no obsolete search result is displayed.
When typing in the "Kiezatlas Objekte suchen" input field quickly, a proper result is displayed always.
It doesn't happen anymore that an obsolete search result is displayed.
This is realized by a logical clock which allows the client to classify server responses as obsolete when already superseded by a more recent request.