Famportal retrieval API: get kiezatlas objects.
The RESTful API allows to retrieve Kiezatlas Geo Objects assigned to a Famportal category.
{{{
GET /famportal/geoobject?category=category-9abc6032-62c9-48a3-af9b-e9cd3931330b-de_DE-1
}}}

The value of the "category" query parameter is the original XML ID of a Famportal category.

The JSON result is an array of Kiezatlas Geo Objects. For the moment only the Geo Object's name is contained in a "name" property. The geo object's coordinates and backlink are pending.

So, this is the very first rudimentary version of the Famportal retrieval API.